requirements.txt
asgiref==3.2.10
Django==3.1
django-debug-toolbar==2.2
djangorestframework==3.11.1
mysqlclient==2.0.1
pytz==2020.1
sqlparse==0.3.1
uwsgi==2.0.19.1
django-allauth==0.44.0
